backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework_simplejwt
drf-yasg